Modify it to a safer script

+++ sync_before_commit.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 set -x
 
+if [[ -n "$1" ]] && [[ "$1" != "rev" ]]; then
+    echo "Wrong Parameter: $1"
+    exit 1
+fi
 
 if [ "$1" == "rev" ]; then
     rsync -a --delete --info=progress2 libgomp/ gcc-8.3.0/libgomp
